+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog  
 
-## Version 0.65.1 - Unreleased
+## Version 0.67.0 - Unreleased
+🍀Added support adding Bar Chart.  
+
+## Version 0.66.0 - 2025-04-18
+🍀Added `IParagraphCollection.Add(string content, int index)` [#986](https://github.com/ShapeCrawler/ShapeCrawler/issues/986)  
 🐞Fixed added picture width [#993](https://github.com/ShapeCrawler/ShapeCrawler/issues/993)  
+🐞Fixed open Google Slides Export [#972](https://github.com/ShapeCrawler/ShapeCrawler/issues/972)  
+🐞Fixed `SetPicture()` [#858](https://github.com/ShapeCrawler/ShapeCrawler/issues/858)  
 
 ## Version 0.65.0 - 2025-04-11
 🍀Added `AsMarkdown()` to get presentation string in Markdown format [(example)](https://github.com/ShapeCrawler/ShapeCrawler/blob/af7042bcdf83e77ecf9b1457415af589cf7bda7a/ShapeCrawler.Examples/Presentation.cs#L7-L12)